Added lots of goodies
Session handeling is implemented, I think.....
All the basic templates are there and the page permissions are there too, so if somone should have acess to a page, it will redirect them to another appropiate page immediately.
Bootstrap, it's not complcated, it just looks scary.
Running all of this off of xammpp locally to test faster, so we may have to change a few things for pluto (specifiaclly the redirects)